tiny

+++ include/am/succinct/fm-index/wavelet_tree_huffman.hpp
@@ -308,7 +308,7 @@ size_t WaveletTreeHuffman<CharT>::select(char_type c, size_t rank) const
 
     for (; walk->parent_; walk = walk->parent_)
     {
-        if ((rank = walk->bit_vector_.Select(rank, bit)) == -1)
+        if ((rank = walk->bit_vector_.Select(rank, bit)) == (size_t)-1)
             return -1;
 
         bit = (walk == walk->parent_->right_);
